modular the secret holder feature for maintainability
